refactor(extra): consolidate shared models
