+++ crates/extra/src/download_link/mod.rs
@@ -3,7 +3,6 @@
 //! 从远程配置加载服务器核心下载链接。
 
 mod manager;
-mod models;
 
+pub use crate::models::{BaseDownloadLinks, DownloadLink, TypeDownloadLinks};
 pub use manager::LinkManager;
-pub use models::{BaseDownloadLinks, DownloadLink, TypeDownloadLinks};
