+++ crates/extra/src/models/download_link.rs
@@ -1,9 +1,10 @@
-//! 下载链接数据模型
+//! 服务器核心下载链接模型。
 
 use serde::{Deserialize, Serialize};
 
-/// 单个下载链接
+/// 单个下载链接。
 #[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
 pub struct DownloadLink {
     pub version: String,
     pub file_name: String,
@@ -16,7 +17,7 @@ impl DownloadLink {
     }
 }
 
-/// 类型下载链接集合
+/// 同一服务器核心类型的版本和下载链接。
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct TypeDownloadLinks {
     pub server_type: String,
@@ -34,11 +35,11 @@ impl TypeDownloadLinks {
     }
 
     pub fn get_link_by_version(&self, version: &str) -> Option<&DownloadLink> {
-        self.links.iter().find(|i| i.version == version)
+        self.links.iter().find(|link| link.version == version)
     }
 }
 
-/// 基础下载链接数据
+/// 所有服务器核心类型的下载链接数据。
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct BaseDownloadLinks {
     pub server_types: Vec<String>,
@@ -55,6 +56,24 @@ impl BaseDownloadLinks {
     }
 
     pub fn get_type_by_name(&self, type_name: &str) -> Option<&TypeDownloadLinks> {
-        self.links.iter().find(|i| i.server_type == type_name)
+        self.links.iter().find(|link| link.server_type == type_name)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::DownloadLink;
+
+    #[test]
+    fn download_link_uses_frontend_field_names() {
+        let value = serde_json::to_value(DownloadLink::new(
+            "1.21.1".to_string(),
+            "server.jar".to_string(),
+            "https://example.invalid/server.jar".to_string(),
+        ))
+        .expect("download link should serialize");
+
+        assert_eq!(value["fileName"], "server.jar");
+        assert!(value.get("file_name").is_none());
     }
 }
