+++ crates/extra/src/models/java.rs
@@ -1,12 +1,37 @@
-//! Java 信息模型
+//! Java 环境信息模型。
 
 use serde::{Deserialize, Serialize};
 
-/// Java 信息
-#[derive(Debug, Clone, Serialize, Deserialize)]
+/// Java 环境信息，用于检测结果和应用设置缓存。
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct JavaInfo {
-    pub version: String,
     pub path: String,
+    pub version: String,
+    pub vendor: String,
+    pub is_64bit: bool,
+    pub major_version: u32,
+    /// Java 安装信息的规则置信度，范围为 0 到 100。
     #[serde(default)]
-    pub major_version: Option<u32>,
+    pub confidence: u8,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::JavaInfo;
+
+    #[test]
+    fn legacy_java_cache_defaults_confidence() {
+        let info: JavaInfo = serde_json::from_str(
+            r#"{
+                "path": "/opt/jdk/bin/java",
+                "version": "21.0.1",
+                "vendor": "OpenJDK",
+                "is_64bit": true,
+                "major_version": 21
+            }"#,
+        )
+        .expect("legacy Java info should remain readable");
+
+        assert_eq!(info.confidence, 0);
+    }
 }
