+++ crates/extra/src/models/mod.rs
@@ -1,14 +1,17 @@
-//! 前端与后端共享的数据模型
+//! `extra` 内跨模块复用的稳定数据模型。
+//!
+//! 领域服务、错误和仅供实现使用的传输结构继续由各自模块维护。
 
 mod app;
+mod app_update;
+mod download_link;
 mod java;
 mod server;
 mod task;
 
-pub use app::AppSettings;
+pub use app::{AppSettings, SettingsGroup, CURRENT_CONFIG_VERSION};
+pub use app_update::{PartialAppSettings, UpdateResult};
+pub use download_link::{BaseDownloadLinks, DownloadLink, TypeDownloadLinks};
 pub use java::JavaInfo;
-pub use server::ServerInstance;
+pub use server::InstanceList;
 pub use task::{TaskProgressResponse, TaskStatus};
-
-// 从 download_link 重新导出
-pub use crate::download_link::{BaseDownloadLinks, DownloadLink, LinkManager, TypeDownloadLinks};
